(1)recover bos download (2) move dsv3_fast_pretrain from env to arg (3) move load_hf_ckpt
