+++ paddleformers/examples/deepseek_v3/load_hf_ckpt.py
